+++ compiler-plugin/src/main/kotlin/tech/mappie/util/Identifiers.kt
@@ -3,6 +3,8 @@ package tech.mappie.util
 import org.jetbrains.kotlin.name.FqName
 import org.jetbrains.kotlin.name.Name
 
+//TODO: contains duplicates
+
 val PACKAGE_MAPPIE_API = FqName("tech.mappie.api")
 
 val PACKAGE_KOTLIN = FqName("kotlin")
@@ -25,6 +27,8 @@ val IDENTIFIER_THROWN_BY_ENUM_ENTRY = Name.identifier("thrownByEnumEntry")
 
 val IDENTIFIER_FROM_PROPERTY = Name.identifier("fromProperty")
 
+val IDENTIFIER_FROM_PROPERTY_NOT_NULL = Name.identifier("fromPropertyNotNull")
+
 val IDENTIFIER_FROM_VALUE = Name.identifier("fromValue")
 
 val IDENTIFIER_FROM_EXPRESSION = Name.identifier("fromExpression")
@@ -34,5 +38,11 @@ val IDENTIFIER_TO = Name.identifier("to")
 val IDENTIFIER_TRANSFORM = Name.identifier("transform")
 
 val IDENTIFIER_VIA = Name.identifier("via")
+
+val PACKAGE_TECH_MAPPIE_API = FqName("tech.mappie.api")
+
+val PACKAGE_TECH_MAPPIE_API_CONFIG = FqName("tech.mappie.api.config")
+
+val CLASS_ID_OBJECT_MAPPING_CONSTRUCTOR = ClassId(PACKAGE_TECH_MAPPIE_API, Name.identifier("ObjectMappingConstructor"))
 
 val IDENTIFIER_LET = Name.identifier("let")