+++ compiler-plugin/src/test/kotlin/tech/mappie/testing/objects/FromExpressionTest.kt
@@ -42,4 +42,54 @@ class FromExpressionTest {
             assertThat(mapper.map(Unit)).isEqualTo(Output(Unit::class.simpleName!!))
         }
     }
+
+    @Test
+    fun `map property fromExpression should succeed with method reference`() {
+        compile(directory) {
+            file("Test.kt",
+                """
+                import tech.mappie.api.ObjectMappie
+                import tech.mappie.testing.objects.FromExpressionTest.*
+
+                class Mapper : ObjectMappie<Int, Output>() {
+                    override fun map(from: Int) = mapping {
+                        Output::value fromExpression Int::toString
+                    }
+                }
+                """
+            )
+        } satisfies {
+            isOk()
+            hasNoMessages()
+
+            val mapper = classLoader
+                .loadObjectMappieClass<Int, Output>("Mapper")
+                .constructors
+                .first()
+                .call()
+
+            assertThat(mapper.map(101)).isEqualTo(Output("101"))
+        }
+    }
+
+    @Test
+    fun `map property fromExpression should fail with method reference with wrong return type`() {
+        compile(directory) {
+            file("Test.kt",
+                """
+                import tech.mappie.api.ObjectMappie
+                import tech.mappie.testing.objects.FromExpressionTest.*
+
+                class Mapper : ObjectMappie<Int, Output>() {
+                    override fun map(from: Int) = mapping {
+                        Output::value fromExpression Int::toInt
+                    }
+                }
+                """
+            )
+        } satisfies {
+            isCompilationError()
+            hasErrorMessage(6, "Target Output::value of type String cannot be assigned from expression of type Int")
+        }
+    }
 }