+++ compiler-plugin/src/test/kotlin/tech/mappie/testing/objects/NestedNullToNonNullPropertyTest.kt
@@ -1,8 +1,10 @@
 package tech.mappie.testing.objects
 
+import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.io.TempDir
 import tech.mappie.testing.compilation.compile
+import tech.mappie.testing.loadObjectMappieClass
 import java.io.File
 
 class NestedNullToNonNullPropertyTest {
@@ -60,6 +62,38 @@ class NestedNullToNonNullPropertyTest {
     }
 
     @Test
+    fun `map object with nested null to non-null explicit fromPropertyNotNull without via should succeed`() {
+        compile(directory) {
+            file("Test.kt",
+                """
+                import tech.mappie.api.ObjectMappie
+                import tech.mappie.testing.objects.NestedNullToNonNullPropertyTest.*
+
+                class Mapper : ObjectMappie<Input, Output>() { 
+                    override fun map(from: Input) = mapping {
+                        to::text fromPropertyNotNull from::text
+                    }
+                }
+
+                object InnerMapper : ObjectMappie<InnerInput, InnerOutput>()
+                """
+            )
+        } satisfies {
+            isOk()
+            hasNoMessages()
+
+            val mapper = classLoader
+                .loadObjectMappieClass<Input, Output>("Mapper")
+                .constructors
+                .first()
+                .call()
+
+            assertThat(mapper.map(Input(InnerInput("test"), 1)))
+                .isEqualTo(Output(InnerOutput("test"), 1))
+        }
+    }
+
+    @Test
     fun `map object with nested null to non-null explicit should fail`() {
         compile(directory) {
             file("Test.kt",
@@ -82,6 +116,38 @@ class NestedNullToNonNullPropertyTest {
                 6,
                 "Target Output::text of type InnerOutput cannot be assigned from from::text via InnerMapper of type InnerOutput?"
             )
+        }
+    }
+
+    @Test
+    fun `map object with nested null to non-null fromPropertyNotNull explicit should succeed`() {
+        compile(directory) {
+            file("Test.kt",
+                """
+                import tech.mappie.api.ObjectMappie
+                import tech.mappie.testing.objects.NestedNullToNonNullPropertyTest.*
+
+                class Mapper : ObjectMappie<Input, Output>() { 
+                    override fun map(from: Input) = mapping {
+                        to::text fromPropertyNotNull from::text via InnerMapper
+                    }
+                }
+
+                object InnerMapper : ObjectMappie<InnerInput, InnerOutput>()
+                """
+            )
+        } satisfies {
+            isOk()
+            hasNoMessages()
+
+            val mapper = classLoader
+                .loadObjectMappieClass<Input, Output>("Mapper")
+                .constructors
+                .first()
+                .call()
+
+            assertThat(mapper.map(Input(InnerInput("test"), 1)))
+                .isEqualTo(Output(InnerOutput("test"), 1))
         }
     }
 }