+++ compiler-plugin/src/test/kotlin/tech/mappie/testing/objects/NullablePropertyToNonNullablePropertyTest.kt
@@ -60,4 +60,33 @@ class NullablePropertyToNonNullablePropertyTest {
             assertThat(mapper.map(Input("value"))).isEqualTo(Output("value"))
         }
     }
+
+    @Test
+    fun `map object null property to non-null property fromPropertyNotNull and transform should succeed`() {
+        compile(directory) {
+            file("Test.kt",
+                """
+                import tech.mappie.api.ObjectMappie
+                import tech.mappie.testing.objects.NullablePropertyToNonNullablePropertyTest.*
+
+                class Mapper : ObjectMappie<Input, Output>() {
+                    override fun map(from: Input) = mapping {
+                        to::value fromPropertyNotNull from::value transform { it + " test" }
+                    }
+                }
+                """
+            )
+        } satisfies {
+            isOk()
+            hasNoMessages()
+
+            val mapper = classLoader
+                .loadObjectMappieClass<Input, Output>("Mapper")
+                .constructors
+                .first()
+                .call()
+
+            assertThat(mapper.map(Input("value"))).isEqualTo(Output("value test"))
+        }
+    }
 }